+++ Library/src/Elements/Movie.java
@@ -1,6 +1,8 @@
-public class Movie extends LibraryItem{
-    private String director;
-    private int duration;
+package Elements;
+
+public class Movie extends LibraryItem {
+    private final String director;
+    private final int duration;
     private static int movieCount = 0;
 
     public Movie(String title, String director, int duration) {
@@ -12,7 +14,7 @@ public class Movie extends LibraryItem{
 
     @Override
     public String toString() {
-        return "Film: " + title + " | Reżyser: " + director + " | Czas trwania: " + duration + " min \n";
+        return "Film: " + getTitle() + " | Reżyser: " + director + " | Czas trwania: " + duration + " min \n";
     }
     public static int getMovieCount() {
         return movieCount;
