+++ Library/src/Exceptions/BookAlreadyBorrowedException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class BookAlreadyBorrowedException extends RuntimeException {
+    public BookAlreadyBorrowedException(String message) {
+        super(message);
+    }
+}
