+++ Library/src/Exceptions/BookIsNotBorrowedException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class BookIsNotBorrowedException extends RuntimeException {
+    public BookIsNotBorrowedException(String message) {
+        super(message);
+    }
+}
