+++ Library/src/Exceptions/ElementNotExistException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class ElementNotExistException extends RuntimeException {
+    public ElementNotExistException(String message) {
+        super(message);
+    }
+}
