+++ Library/src/Exceptions/ZeroAvailableElementException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class ZeroAvailableElementException extends RuntimeException {
+    public ZeroAvailableElementException(String message) {
+        super(message);
+    }
+}
