+++ Library/src/Exceptions/ZeroBorrowedElementException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class ZeroBorrowedElementException extends RuntimeException {
+    public ZeroBorrowedElementException(String message) {
+        super(message);
+    }
+}
